refactor(store): put model package inside postgres
