+++ store/postgres/provider_repository.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/odpf/guardian/core/provider"
 	"github.com/odpf/guardian/domain"
-	"github.com/odpf/guardian/store/model"
+	"github.com/odpf/guardian/store/postgres/model"
 	"gorm.io/gorm"
 )
